requirements.txt
# Core dependencies for LAJ-Gherkin
openai==1.88.0
python-dotenv==1.1.0
PyYAML==6.0.2

# Data analysis and processing
pandas==2.3.2
numpy==2.2.3
openpyxl==3.1.5

# Visualization (for notebooks)
matplotlib==3.10.6
seaborn==0.13.2

# Jupyter support (for analysis notebooks)
ipykernel==6.30.1

# Utilities
tabulate==0.9.0
tqdm==4.67.1
